refactor + update pastWidget action function
