+++ app/client/src/sagas/CanvasSagas/DraggingCanvasSagas.ts
@@ -26,31 +26,22 @@ import type {
   CanvasWidgetsReduxState,
   FlattenedWidgetProps,
 } from "reducers/entityReducers/canvasWidgetsReducer";
-import type { DragDetails } from "reducers/uiReducers/dragResizeReducer";
 import type { MainCanvasReduxState } from "reducers/uiReducers/mainCanvasReducer";
 import { all, call, put, select, takeLatest } from "redux-saga/effects";
-import { addBuildingBlockToApplication } from "sagas/BuildingBlocksSagas";
+import { addAndMoveBuildingBlockToCanvasSaga } from "sagas/BuildingBlockAdditionSagas";
 import { getUpdateDslAfterCreatingChild } from "sagas/WidgetAdditionSagas";
 import {
   executeWidgetBlueprintBeforeOperations,
   traverseTreeAndExecuteBlueprintChildOperations,
 } from "sagas/WidgetBlueprintSagas";
-import {
-  getDragDetails,
-  getWidget,
-  getWidgetByName,
-  getWidgets,
-  getWidgetsMeta,
-} from "sagas/selectors";
+import { getWidget, getWidgets, getWidgetsMeta } from "sagas/selectors";
 import {
   getCanvasWidth,
-  getCurrentApplicationId,
   getIsAutoLayoutMobileBreakPoint,
   getMainCanvasProps,
   getOccupiedSpacesSelectorForContainer,
 } from "selectors/editorSelectors";
 import { getLayoutSystemType } from "selectors/layoutSystemSelectors";
-import { initiateBuildingBlockDropEvent } from "utils/buildingBlockUtils";
 import { collisionCheckPostReflow } from "utils/reflowHookUtils";
 import type { WidgetProps } from "widgets/BaseWidget";
 
@@ -150,52 +141,6 @@ const getBottomMostRowAfterMove = (
   return widgetBottomRow;
 };
 
-function* addBuildingBlockAndMoveWidgetsSaga(
-  actionPayload: ReduxAction<{
-    newWidget: WidgetAddChild;
-    draggedBlocksToUpdate: WidgetDraggingUpdateParams[];
-    canvasId: string;
-  }>,
-) {
-  const applicationId: string = yield select(getCurrentApplicationId);
-  const workspaceId: string = yield select(getCurrentApplicationId);
-  const dragDetails: DragDetails = yield select(getDragDetails);
-  const buildingblockName = dragDetails.newWidget.displayName;
-  const skeletonWidgetName = `loading_${buildingblockName
-    .toLowerCase()
-    .replace(/ /g, "_")}`;
-
-  yield call(addWidgetAndMoveWidgetsSaga, {
-    ...actionPayload,
-    payload: {
-      ...actionPayload.payload,
-      // so that the skeleton loader does not get included when the users uses the undo/redo
-      shouldReplay: false,
-      newWidget: {
-        ...actionPayload.payload.newWidget,
-        type: "SKELETON_WIDGET",
-        widgetName: skeletonWidgetName,
-        widgetId: MAIN_CONTAINER_WIDGET_ID,
-      },
-    },
-  });
-  yield call(initiateBuildingBlockDropEvent, {
-    applicationId,
-    workspaceId,
-    buildingblockName,
-  });
-
-  const skeletonWidget: FlattenedWidgetProps = yield select(
-    getWidgetByName,
-    skeletonWidgetName,
-  );
-  yield call(
-    addBuildingBlockToApplication,
-    actionPayload.payload.newWidget,
-    skeletonWidget.widgetId,
-  );
-}
-
 function* addAndMoveUIEntitySaga(
   actionPayload: ReduxAction<{
     newWidget: WidgetAddChild;
@@ -204,7 +149,7 @@ function* addAndMoveUIEntitySaga(
   }>,
 ) {
   if (actionPayload.payload.newWidget.type === BUILDING_BLOCK_EXPLORER_TYPE) {
-    yield call(addBuildingBlockAndMoveWidgetsSaga, actionPayload);
+    yield call(addAndMoveBuildingBlockToCanvasSaga, actionPayload);
   } else {
     yield call(addWidgetAndMoveWidgetsSaga, actionPayload);
   }
